requirements.txt
#
#
# These materials contain confidential information and
# trade secrets of Dynatrace LLC.  You shall
# maintain the materials as confidential and shall not
# disclose its contents to any third party except as may
# be required by law or regulation.  Use, disclosure,
# or reproduction is prohibited without the prior express
# written permission of Dynatrace LLC.
# 
# All Compuware products listed within the materials are
# trademarks of Dynatrace LLC.  All other company
# or product names are trademarks of their respective owners.
# 
# Copyright (c) 2024 Dynatrace LLC.  All rights reserved.
#
#
snowflake>=1.0.3 # IMPORTANT: Requires Python >=3.9 and <3.12 (due to snowflake-legacy dependency)
pandas
toml
tzlocal
#%DEV:
streamlit
snowflake-snowpark-python
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
uuid
jsonstrip
numpy
pytest
inflect
markdown2
weasyprint
pyyaml
pylint
flake8
pylint
#%:DEV
